+++ sycl/include/sycl/handler.hpp
@@ -33,11 +33,11 @@
 #include <sycl/ext/oneapi/device_global/device_global.hpp>
 #include <sycl/ext/oneapi/device_global/properties.hpp>
 #include <sycl/ext/oneapi/experimental/graph.hpp>
+#include <sycl/ext/oneapi/experimental/raw_kernel_arg.hpp>
 #include <sycl/ext/oneapi/experimental/use_root_sync_prop.hpp>
 #include <sycl/ext/oneapi/experimental/virtual_functions.hpp>
 #include <sycl/ext/oneapi/kernel_properties/properties.hpp>
 #include <sycl/ext/oneapi/properties/properties.hpp>
-#include <sycl/ext/oneapi/raw_kernel_arg.hpp>
 #include <sycl/group.hpp>
 #include <sycl/id.hpp>
 #include <sycl/item.hpp>
@@ -497,7 +497,8 @@ private:
     return Storage;
   }
 
-  void *storeRawArg(const sycl::ext::oneapi::raw_kernel_arg &RKA) {
+  void *
+  storeRawArg(const sycl::ext::oneapi::experimental::raw_kernel_arg &RKA) {
     CGData.MArgsStorage.emplace_back(RKA.MArgSize);
     void *Storage = static_cast<void *>(CGData.MArgsStorage.back().data());
     std::memcpy(Storage, RKA.MArgData, RKA.MArgSize);
@@ -730,7 +731,8 @@ private:
   }
 
   // setArgHelper for the raw_kernel_arg extension type.
-  void setArgHelper(int ArgIndex, sycl::ext::oneapi::raw_kernel_arg &&Arg) {
+  void setArgHelper(int ArgIndex,
+                    sycl::ext::oneapi::experimental::raw_kernel_arg &&Arg) {
     auto StoredArg = storeRawArg(Arg);
     MArgs.emplace_back(detail::kernel_param_kind_t::kind_std_layout, StoredArg,
                        Arg.MArgSize, ArgIndex);
@@ -2032,7 +2034,7 @@ public:
   }
 
   // set_arg for the raw_kernel_arg extension type.
-  void set_arg(int argIndex, ext::oneapi::raw_kernel_arg &&Arg) {
+  void set_arg(int argIndex, ext::oneapi::experimental::raw_kernel_arg &&Arg) {
     setArgHelper(argIndex, std::move(Arg));
   }
